Simplify resolution of module data (ex-jrtfs) and lib/ct.sym
All our data is available through the classpath dependency of
jdk.compiler.home, so we can simplify the lookup of this data.

We also only have one PlatformProvider (JDKPlatformProvider), so we
don't have to look it up via the Service api.

Cut down the logic to the minimum we need.

This also fixes a bug in connection with GraalVM native-image.